requirements.txt
flask==2.2.3
werkzeug==2.2.3
google-generativeai==0.3.1
